+++ test/test-utils/index.ts
@@ -1,2 +1,3 @@
 export * from './test-utils';
 export * from './wrappers';
+export * from './utilities';
