+++ test/utils/test-utils.ts
@@ -53,58 +53,3 @@ export const mockStateEventImplementation = (events: MatrixEvent[]) => {
         return Array.from(stateMap.get(eventType)?.values() || []);
     };
 };
-
-export const mkRoom = (client: MatrixClient, roomId: string, rooms?: ReturnType<typeof mkStubRoom>[]) => {
-    const room = mkStubRoom(roomId, roomId, client);
-    room.currentState.getStateEvents.mockImplementation(mockStateEventImplementation([]));
-    rooms?.push(room);
-    return room;
-};
-
-/**
- * Upserts given events into room.currentState
- * @param room
- * @param events
- */
-export const upsertRoomStateEvents = (room: Room, events: MatrixEvent[]): void => {
-    const eventsMap = events.reduce((acc, event) => {
-        const eventType = event.getType();
-        if (!acc.has(eventType)) {
-            acc.set(eventType, new Map());
-        }
-        acc.get(eventType).set(event.getStateKey(), event);
-        return acc;
-    }, room.currentState.events || new Map<string, Map<string, MatrixEvent>>());
-
-    room.currentState.events = eventsMap;
-};
-
-export const mkSpace = (
-    client: MatrixClient,
-    spaceId: string,
-    rooms?: ReturnType<typeof mkStubRoom>[],
-    children: string[] = [],
-) => {
-    const space = mkRoom(client, spaceId, rooms);
-    space.isSpaceRoom.mockReturnValue(true);
-    space.currentState.getStateEvents.mockImplementation(mockStateEventImplementation(children.map(roomId =>
-        mkEvent({
-            event: true,
-            type: EventType.SpaceChild,
-            room: spaceId,
-            user: "@user:server",
-            skey: roomId,
-            content: { via: [] },
-            ts: Date.now(),
-        }),
-    )));
-    return space;
-};
-
-export const emitPromise = (e: EventEmitter, k: string | symbol) => new Promise(r => e.once(k, r));
-
-const findByAttr = (attr: string) => (component: ReactWrapper, value: string) => component.find(`[${attr}="${value}"]`);
-export const findByTestId = findByAttr('data-test-id');
-export const findById = findByAttr('id');
-
-export const flushPromises = async () => await new Promise(resolve => setTimeout(resolve));
